// types/IResponse.ts
export interface FileWithMetadata {
  name: any;
  imageType: string;
  file: File;
  imageData: string;
}
// ...
export interface FormValuesWork {
  studentId?: string;                   // Student ID
  prefix?: string;                      // Prefix (Mr., Ms., Dr., etc.)
  firstName?: string;                   // Full name of the student
  nickname?: string;                    // Nickname of the student
  graduate?: string;                    // Graduation year or status
  branch?: string;                      // Branch of study
  workName?: string;                    // Special work name
  organizationName?: string;            // Organization name
  organizationPhone?: string;           // Organization phone number
  workDescription?: string;             // Description of the special work
  activityDate?: string;                // Date of the activity
  hours?: number;                       // Number of hours worked
  workType?: string;                    // Type of work (e.g., paid, volunteer)
  compensation?: number;                // Compensation received, if any
  workDates?: string;                   // Work dates
  workTime?: string;  
  Specialwork ?: string;                  // Work time
  uploadSpecialwork?: FileList | File[]; // File(s) for special work upload
  specialworkId?:string
  id?:string
  
  uploadVolunteer?:FileList| File

}